EMP-408 create start update test script

+++ test/start_update.sh
@@ -11,33 +11,38 @@ set -eu
 # shellcheck source=test/include/chroot_env.sh
 . "test/include/chroot_env.sh"
 
+# Common directory variables
+SBINDIR="/sbin"
+
 ARM_EMU_BIN="${ARM_EMU_BIN:-}"
 
 CWD="$(pwd)"
 
 SYSTEM_UPDATE_DIR="${SYSTEM_UPDATE_DIR:-/etc/system_update}"
-UPDATE_EXCLUDE_LIST_FILE="jedi_update_exclude_list.txt"
+START_UPDATE_COMMAND="${SBINDIR}/start_update.sh"
 UPDATE_ROOTFS_SOURCE="/tmp/update_source"
 TARGET_STORAGE_DEVICE=""
 
 JEDI_PARTITION_TABLE_FILE_NAME="config/jedi_emmc_sfdisk.table"
-UPDATE_FILES_COMMAND="${SYSTEM_UPDATE_DIR}.d/50_update_files.sh"
-
-ULTIMAKER_VERSION_FILE="/etc/ultimaker_version"
-DEBIAN_VERSION_FILE="/etc/debian_version"
+UPDATE_EXCLUDE_LIST_FILE="config/jedi_update_exclude_list.txt"
 
 STORAGE_DEVICE_IMG="storage_device.img"
 BYTES_PER_SECTOR="512"
 STORAGE_DEVICE_SIZE="7553024" # sectors, about 3.6 GiB
 
+TEST_UPDATE_ROOTFS_FILE="test/test_rootfs.tar.xz"
+TEMP_TEST_UPDATE_ROOTFS_FILE="rootfs-v1.2.3.tar.xz"
+
 TEST_UPDATE_ROOTFS_FILE="${CWD}/test/test_rootfs.tar.xz"
 TEST_OUTPUT_FILE="$(mktemp -d)/test_results_$(basename "${0%.sh}").txt"
 
 NAME_TEMPLATE_TOOLBOX="um-update-toolbox"
-NAME_TEMPLATE_WORKDIR="update_files_workdir"
+NAME_TEMPLATE_UPDATEROOT="temp_update_mount"
+NAME_TEMPLATE_WORKDIR="start_update_workdir"
 
 toolbox_image=""
 toolbox_root_dir=""
+update_mount=""
 work_dir=""
 
 exit_on_failure=false
@@ -79,15 +84,17 @@ create_dummy_storage_device()
 
 setup()
 {
-    toolbox_root_dir="$(mktemp -d -t "${NAME_TEMPLATE_WORKDIR}.XXXXXX")"
+    toolbox_root_dir="$(mktemp -d -t "${NAME_TEMPLATE_TOOLBOX}.XXXXXX")"
 
     setup_chroot_env "${toolbox_image}" "${toolbox_root_dir}"
 
-    mkdir -p "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}"
-    tar xvfJ "${TEST_UPDATE_ROOTFS_FILE}" -C "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}" \
+    mkdir -p "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}"
+    tar -xJvf "${TEST_UPDATE_ROOTFS_FILE}" -C "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}" \
         > /dev/null 2> /dev/null
 
-    work_dir="$(mktemp -d -t "${NAME_TEMPLATE_TOOLBOX}.XXXXXX")"
+    update_mount="$(mktemp -d -t "${NAME_TEMPLATE_UPDATEROOT}.XXXXXX")"
+
+    work_dir="$(mktemp -d -t "${NAME_TEMPLATE_WORKDIR}.XXXXXX")"
     cd "${work_dir}"
 
     create_dummy_storage_device
@@ -97,23 +104,28 @@ teardown()
 {
     teardown_chroot_env "${toolbox_root_dir}"
 
+    cd "${CWD}"
+
     if [ -b "${TARGET_STORAGE_DEVICE}" ]; then
         losetup -d "${TARGET_STORAGE_DEVICE}"
         TARGET_STORAGE_DEVICE=""
     fi
 
-    if [ -d "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}" ] && [ -z "${toolbox_root_dir##*um-update-toolbox*}" ]; then
-        rm -rf "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}"
+    if [ -d "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}" ] && \
+       [ -z "${toolbox_root_dir##*${NAME_TEMPLATE_TOOLBOX}*}" ]; then
+        rm -rf "${toolbox_root_dir:?}/${UPDATE_ROOTFS_SOURCE}"
     fi
 
-    cd "${CWD}"
-
-    if [ -d "${work_dir}" ] && [ -z "${work_dir##*update_files_workdir*}" ]; then
-        rm -r "${work_dir}"
+    if [ -d "${toolbox_root_dir}" ] && [ -z "${toolbox_root_dir##*${NAME_TEMPLATE_TOOLBOX}*}" ]; then
+        rm -rf "${toolbox_root_dir}"
     fi
 
-    if [ -d "${toolbox_root_dir}" ] && [ -z "${toolbox_root_dir##*um-update-toolbox*}" ]; then
-        rm -r "${toolbox_root_dir}"
+    if [ -d "${update_mount}" ] && [ -z "${update_mount##*${NAME_TEMPLATE_UPDATEROOT}*}" ]; then
+        rm -rf "${update_mount}"
+    fi
+
+    if [ -d "${work_dir}" ] && [ -z "${work_dir##*${NAME_TEMPLATE_WORKDIR}*}" ]; then
+        rm -rf "${work_dir}"
     fi
 }
 
@@ -123,7 +135,7 @@ failure_exit()
     echo "When finished, the following is needed to cleanup!"
     echo "  sudo sh -c '\\"
     echo "    losetup -d '${TARGET_STORAGE_DEVICE}' && \\"
-    echo "    rm -rf '${work_dir:?}/*' && \\"
+    echo "    rm -rf '${work_dir}/*' && \\"
 
     failure_exit_chroot_env
 
@@ -165,60 +177,49 @@ run_test()
     teardown
 }
 
-test_missing_argument_update_rootfs_source_nok()
+test_rsync_ignore_file_not_found_nok()
 {
-    chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" -s '' -d "${TARGET_STORAGE_DEVICE}" || return 0
+    cp "${TEST_UPDATE_ROOTFS_FILE}" "${update_mount}/${TEMP_TEST_UPDATE_ROOTFS_FILE}"
+    rm "${toolbox_root_dir:?}/${SYSTEM_UPDATE_DIR:?}/"*_exclude_list.txt
+    "${toolbox_root_dir}/${START_UPDATE_COMMAND}" "${toolbox_root_dir}" "${update_mount}" "${TARGET_STORAGE_DEVICE}" || return 0
 }
 
-test_update_rootfs_source_not_a_directory_nok()
+test_partition_table_not_found_nok()
 {
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=/tmp/not_existing_dir"
-
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 0
+    cp "${TEST_UPDATE_ROOTFS_FILE}" "${update_mount}/${TEMP_TEST_UPDATE_ROOTFS_FILE}"
+    rm "${toolbox_root_dir:?}/${SYSTEM_UPDATE_DIR:?}/"*.table
+    "${toolbox_root_dir}/${START_UPDATE_COMMAND}" "${toolbox_root_dir}" "${update_mount}" "${TARGET_STORAGE_DEVICE}" || return 0
 }
 
-test_invalid_block_device_nok()
+test_update_rootfs_corrupt_nok()
 {
-    target_device="/dev/loop100"
-    chroot "${toolbox_root_dir}" "${UPDATE_FILES_COMMAND}" -s "${UPDATE_ROOTFS_SOURCE}" -d "${target_device}" || return 0
+    cp "${TEST_UPDATE_ROOTFS_FILE}" "${update_mount}/${TEMP_TEST_UPDATE_ROOTFS_FILE}"
+    echo "Append this data to corrupt the archive" >> "${update_mount}/${TEMP_TEST_UPDATE_ROOTFS_FILE}"
+    "${toolbox_root_dir}/${START_UPDATE_COMMAND}" "${toolbox_root_dir}" "${update_mount}" "${TARGET_STORAGE_DEVICE}" || return 0
 }
 
-test_no_debian_distribution_found_nok()
+test_multiple_update_rootfs_files_nok()
 {
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE}"
-
-    unlink "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}${DEBIAN_VERSION_FILE}"
-
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 0
+    cp "${TEST_UPDATE_ROOTFS_FILE}" "${update_mount}/${TEMP_TEST_UPDATE_ROOTFS_FILE}"
+    cp "${update_mount}/${TEMP_TEST_UPDATE_ROOTFS_FILE}" \
+        "${update_mount}/rootfs-v2.1.0.tar.xz"
+    "${toolbox_root_dir}/${START_UPDATE_COMMAND}" "${toolbox_root_dir}" "${update_mount}" "${TARGET_STORAGE_DEVICE}" || return 0
 }
 
-test_no_ultimaker_software_found_nok()
+test_successful_update_ok()
 {
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE}"
+    cp "${TEST_UPDATE_ROOTFS_FILE}" "${update_mount}/${TEMP_TEST_UPDATE_ROOTFS_FILE}"
+    "${toolbox_root_dir}/${START_UPDATE_COMMAND}" "${toolbox_root_dir}" "${update_mount}" "${TARGET_STORAGE_DEVICE}" || return 1
 
-    unlink "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}${ULTIMAKER_VERSION_FILE}"
+    update_target="$(mktemp -d)"
+    mount -t auto -v "${TARGET_STORAGE_DEVICE}p2" "${update_target}"
 
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 0
-}
-
-test_update_files_ok()
-{
-    chroot_environment="TARGET_STORAGE_DEVICE=${TARGET_STORAGE_DEVICE}"
-    chroot_environment="${chroot_environment} UPDATE_ROOTFS_SOURCE=${UPDATE_ROOTFS_SOURCE}"
-    chroot "${toolbox_root_dir}" /bin/sh -c "${chroot_environment} ${UPDATE_FILES_COMMAND}" || return 1
-
-    update_target="$(mktemp -d -t "tmp_update_target.XXXXXX")"
-    mount -t auto -v "${TARGET_STORAGE_DEVICE}p2" "${update_target}" || return 1
-
-    rsync --exclude-from "${CWD}/config/${UPDATE_EXCLUDE_LIST_FILE}" -c -a -x --dry-run \
-        "${toolbox_root_dir}${UPDATE_ROOTFS_SOURCE}/" "${update_target}/" || return 1
+    rsync --exclude-from "${CWD}/${UPDATE_EXCLUDE_LIST_FILE}" -c -a -x --dry-run \
+        "${toolbox_root_dir}/${UPDATE_ROOTFS_SOURCE}/" "${update_target}/" || return 1
 
     umount "${update_target}"
 
-    if [ -z "${update_target##*tmp_update_target*}" ]; then
+    if [ -z "${update_target##*/tmp/*}" ]; then
         rm -rf "${update_target}"
     fi
 }
@@ -275,12 +276,11 @@ fi
 
 trap cleanup EXIT
 
-run_test test_missing_argument_update_rootfs_source_nok
-run_test test_update_rootfs_source_not_a_directory_nok
-run_test test_invalid_block_device_nok
-run_test test_no_debian_distribution_found_nok
-run_test test_no_ultimaker_software_found_nok
-run_test test_update_files_ok
+run_test test_partition_table_not_found_nok
+run_test test_rsync_ignore_file_not_found_nok
+run_test test_update_rootfs_corrupt_nok
+run_test test_multiple_update_rootfs_files_nok
+run_test test_successful_update_ok
 
 echo "________________________________________________________________________________"
 echo "Test results '${TEST_OUTPUT_FILE}':"
